# PERF/CLN: float-to-int casting