@jorisvandenbossche 2020-01-29
There should be a similar case in `arrays/integer.py` that you can update, I think