@jreback 2020-01-31
more cases in pandas/core/dtypes/cast.py no?