@jbrockmendel 2020-02-01
> more cases in pandas/core/dtypes/cast.py no?

Just did a pass through the file, didnt find any.

> also cast_scalar_indexer really belongs in pandas/core/dtypes/cast.py as well (common.py is a hodge podge)

Probably.  I've got a lot of refactor-esque PRs/branches going, will move this after the current batch.